# fix: reply hasHeader behavior