--- a/lib/reply.js
+++ b/lib/reply.js
@@ -209,7 +209,10 @@ Reply.prototype.getHeaders = function () {
 }
 
 Reply.prototype.hasHeader = function (key) {
-  return this[kReplyHeaders][key.toLowerCase()] !== undefined
+  if (this[kReplyHeaders][key.toLowerCase()] !== undefined) {
+    return true
+  }
+  return this.raw.hasHeader(key)
 }
 
 Reply.prototype.removeHeader = function (key) {