--- a/lib/reply.js
+++ b/lib/reply.js
@@ -209,7 +209,8 @@ Reply.prototype.getHeaders = function () {
 }
 
 Reply.prototype.hasHeader = function (key) {
-  if (this[kReplyHeaders][key.toLowerCase()] !== undefined) {
+  key = key.toLowerCase()
+  if (this[kReplyHeaders][key] !== undefined) {
     return true
   }
   return this.raw.hasHeader(key)