@zekth 2021-04-23
Benchmarks added.

Side note, we can also use the branch free approach like:
```js
Reply.prototype.hasHeader = function (key) {
  key = key.toLowerCase()
  return this[kReplyHeaders][key] !== undefined || this.raw.hasHeader(key)
}
```